account ids

+++ app/src/main/java/com/example/mobfinalmenuplusnavbar/ui/main/RecyclerMainViewAdapter.java
@@ -17,7 +17,7 @@ import java.util.ArrayList;
 public class RecyclerMainViewAdapter extends RecyclerView.Adapter<RecyclerMainViewAdapter.ViewHolder> {
 
     private static final String TAG = "RecyclerViewAdapter";
-    private ArrayList<Integer> cardRecyclerIds = new ArrayList<>();
+    private ArrayList<Long> cardRecyclerIds = new ArrayList<>();
     private ArrayList<Integer> cardRecyclerLogos = new ArrayList<>();
     private ArrayList<String> cardRecyclerBankNames = new ArrayList<>();
     private ArrayList<String> cardRecyclerCashes = new ArrayList<>();
@@ -25,7 +25,7 @@ public class RecyclerMainViewAdapter extends RecyclerView.Adapter<RecyclerMainVi
 
     public RecyclerMainViewAdapter(
             Context cardRecyclerContext,
-            ArrayList<Integer> cardRecyclerIds,
+            ArrayList<Long> cardRecyclerIds,
             ArrayList<Integer> cardRecyclerLogos,
             ArrayList<String> cardRecyclerBankNames,
             ArrayList<String> cardRecyclerCashes
